Provide parameter to disable use of signalfx repo
In a large managed environment, common practice is to block direct
Internet access from most hosts and to store all packages in locally
curated repositories.

This provides a parameter to block the default behavior of installing a
remote repository. The user is then responsible for putting appropriate
packages in a local repository that is accessible to the client hosts.